added test command to automate most testing

+++ built/commands/play.js
@@ -63,8 +63,10 @@ const playCommand = {
                     .catch((err) => {
                     warnCustom(message, `Error Queuing Song, Please Try Again.\n\nError:\n${err.message}`, this.name);
                 });
-                message.delete();
-                message.deleted = true;
+                if (!message.deleted) {
+                    message.delete();
+                    message.deleted = true;
+                }
             }
         });
     },
