+++ built/commands/playnext.js
@@ -79,8 +79,10 @@ const playNextCommand = {
                     textChannel: channel,
                     position: 1,
                 });
-                message.delete();
-                message.deleted = true;
+                if (!message.deleted) {
+                    message.delete();
+                    message.deleted = true;
+                }
             }
         });
     },
