+++ src/commands/say.ts
@@ -33,8 +33,10 @@ const sayCommand: Command = {
         }
 
         channel.send(argsString);
-        message.delete();
-        message.deleted = true;
+        if (!message.deleted) {
+            message.delete();
+            message.deleted = true;
+        }
     },
 };
 //#endregion
